refactor(cli): move vk commands to sdk
- [x] Refactor sdk structure for verifying keys
- [x] Move extractVkToFile to sdk
- [x] Move checkVerifyingKeys to sdk
- [x] Move setVerifyingKeys to sdk
