+++ packages/cli/ts/commands/index.ts
@@ -3,12 +3,9 @@ export { deployPoll } from "./deployPoll";
 export { joinPoll, isJoinedUser } from "./joinPoll";
 export { deployVkRegistryContract } from "./deployVkRegistry";
 export { publish, publishBatch } from "./publish";
-export { setVerifyingKeysCli } from "./setVerifyingKeys";
 export { showContracts } from "./showContracts";
 export { timeTravel } from "./timeTravel";
 export { genProofs } from "./genProofs";
 export { fundWallet } from "./fundWallet";
 export { proveOnChain } from "./proveOnChain";
-export { checkVerifyingKeys } from "./checkVerifyingKeys";
 export { genLocalState } from "./genLocalState";
-export { extractVkToFile } from "./extractVkToFile";
