+++ packages/sdk/ts/keys/setVerifyingKeys.ts
@@ -1,0 +1,114 @@
+import { EMode, IVerifyingKeyStruct, VkRegistry__factory as VkRegistryFactory } from "maci-contracts";
+import { genPollJoinedVkSig, genPollJoiningVkSig, genProcessVkSig, genTallyVkSig } from "maci-core";
+
+import type { ISetVerifyingKeysArgs } from "./types";
+
+import { contractExists } from "../utils";
+
+import { compareVks } from "./utils";
+
+/**
+ * Set the verifying keys on the contract
+ * @param args - The arguments for the setVerifyingKeys function
+ */
+export const setVerifyingKeys = async ({
+  pollJoiningVk,
+  pollJoinedVk,
+  processMessagesVk,
+  tallyVotesVk,
+  stateTreeDepth,
+  intStateTreeDepth,
+  voteOptionTreeDepth,
+  messageBatchSize,
+  vkRegistryAddress,
+  signer,
+  mode,
+}: ISetVerifyingKeysArgs): Promise<void> => {
+  // validate args
+  if (stateTreeDepth < 1 || intStateTreeDepth < 1 || voteOptionTreeDepth < 1 || messageBatchSize < 1) {
+    throw new Error("Invalid depth or batch size parameters");
+  }
+
+  if (stateTreeDepth < intStateTreeDepth) {
+    throw new Error("Invalid state tree depth or intermediate state tree depth");
+  }
+
+  const vkRegistryContractExists = await contractExists(signer.provider!, vkRegistryAddress);
+  // ensure we have a contract deployed at the provided address
+  if (!vkRegistryContractExists) {
+    throw new Error(`A VkRegistry contract is not deployed at ${vkRegistryAddress}`);
+  }
+
+  // connect to VkRegistry contract
+  const vkRegistryContract = VkRegistryFactory.connect(vkRegistryAddress, signer);
+
+  // check if the poll vk was already set
+  const pollJoiningVkSig = genPollJoiningVkSig(stateTreeDepth);
+
+  if (await vkRegistryContract.isPollJoiningVkSet(pollJoiningVkSig)) {
+    throw new Error("This poll verifying key is already set in the contract");
+  }
+
+  // check if the poll vk was already set
+  const pollJoinedVkSig = genPollJoinedVkSig(stateTreeDepth);
+
+  if (await vkRegistryContract.isPollJoinedVkSet(pollJoinedVkSig)) {
+    throw new Error("This poll verifying key is already set in the contract");
+  }
+
+  // check if the process messages vk was already set
+  const processVkSig = genProcessVkSig(stateTreeDepth, voteOptionTreeDepth, messageBatchSize);
+
+  if (await vkRegistryContract.isProcessVkSet(processVkSig, EMode.QV)) {
+    throw new Error("This process verifying key is already set in the contract");
+  }
+
+  // do the same for the tally votes vk
+  const tallyVkSig = genTallyVkSig(stateTreeDepth, intStateTreeDepth, voteOptionTreeDepth);
+
+  if (await vkRegistryContract.isTallyVkSet(tallyVkSig, mode)) {
+    throw new Error("This tally verifying key is already set in the contract");
+  }
+
+  // set them onchain
+  const tx = await vkRegistryContract.setVerifyingKeysBatch(
+    stateTreeDepth,
+    intStateTreeDepth,
+    voteOptionTreeDepth,
+    messageBatchSize,
+    [mode],
+    pollJoiningVk.asContractParam() as IVerifyingKeyStruct,
+    pollJoinedVk.asContractParam() as IVerifyingKeyStruct,
+    [processMessagesVk.asContractParam() as IVerifyingKeyStruct],
+    [tallyVotesVk.asContractParam() as IVerifyingKeyStruct],
+  );
+
+  const receipt = await tx.wait();
+
+  if (receipt?.status !== 1) {
+    throw new Error("Set verifying keys transaction failed");
+  }
+
+  const [pollJoiningVkOnChain, pollJoinedVkOnChain, processVkOnChain, tallyVkOnChain] = await Promise.all([
+    vkRegistryContract.getPollJoiningVk(stateTreeDepth),
+    vkRegistryContract.getPollJoinedVk(stateTreeDepth),
+    vkRegistryContract.getProcessVk(stateTreeDepth, voteOptionTreeDepth, messageBatchSize, mode),
+    vkRegistryContract.getTallyVk(stateTreeDepth, intStateTreeDepth, voteOptionTreeDepth, mode),
+  ]);
+
+  if (!compareVks(pollJoiningVkOnChain, pollJoiningVk)) {
+    throw new Error("pollJoiningVk mismatch");
+  }
+
+  if (!compareVks(pollJoinedVkOnChain, pollJoinedVk)) {
+    throw new Error("pollJoinedVk mismatch");
+  }
+
+  if (!compareVks(processVkOnChain, processMessagesVk)) {
+    throw new Error("processVk mismatch");
+  }
+
+  if (!compareVks(tallyVkOnChain, tallyVotesVk)) {
+    throw new Error("tallyVk mismatch");
+  }
+};
